Add hint with qestion mark and popup
